kiosk/requirements-test.txt
# Test dependencies for Smart Hotel Kiosk
# Install with: pip install -r requirements-test.txt

-r requirements.txt

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
factory-boy>=3.3.0
faker>=22.0.0

# HTTP mocking
responses>=0.24.0
requests-mock>=1.11.0

# Code quality
coverage>=7.3.0
